requirements.txt
# Career Assistant AI Agent - Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
openai>=1.12.0
google-generativeai>=0.8.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
httpx>=0.26.0
aiofiles>=23.2.0
pytest>=8.0.0
